requirements.txt
pandas>=1.5.0
numpy>=1.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.1.0
nltk>=3.7
pathlib
accelerate
